Moved functions/operators with `lhs: Scalar` into separate ‘Scalar.swift’ file
